Created a simple ObjectManager and put it to use. Changed the deltaTime type being passed in Update and Move functions to double to increase accuracy and avoid implicit conversion.

+++ KF6017/GameObject.cpp
@@ -23,7 +23,7 @@ void GameObject::Initialise(Vector2D position, float angle, float scale)
 	SetActive(true);
 }
 
-ErrorType GameObject::Update(float deltaTime)
+ErrorType GameObject::Update(double deltaTime)
 {
 	if (!IsActive())
 		return SUCCESS;
